refactor: 프로덕션 코드 패키지 리팩터링
