+++ backend/src/main/java/com/woowacourse/zzimkkong/controller/AuthController.java
@@ -1,7 +1,7 @@
 package com.woowacourse.zzimkkong.controller;
 
-import com.woowacourse.zzimkkong.dto.LoginRequest;
-import com.woowacourse.zzimkkong.dto.TokenResponse;
+import com.woowacourse.zzimkkong.dto.member.LoginRequest;
+import com.woowacourse.zzimkkong.dto.member.TokenResponse;
 import com.woowacourse.zzimkkong.service.AuthService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
