+++ backend/src/main/java/com/woowacourse/zzimkkong/dto/member/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.zzimkkong.dto;
+package com.woowacourse.zzimkkong.dto.member;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
@@ -7,26 +7,21 @@ import javax.validation.constraints.Pattern;
 
 import static com.woowacourse.zzimkkong.dto.Validator.*;
 
-public class MemberSaveRequest {
+public class LoginRequest {
     @NotBlank(message = EMPTY_MESSAGE)
     @Email(message = EMAIL_MESSAGE)
     private String email;
 
     @NotNull(message = EMPTY_MESSAGE)
-    @Pattern(regexp = "^(?=.*[0-9])(?=.*[a-zA-Z]).{8,20}$", message = MEMBER_PASSWORD_MESSAGE)
+    @Pattern(regexp = MEMBER_PASSWORD_FORMAT, message = MEMBER_PASSWORD_MESSAGE)
     private String password;
 
-    @NotNull(message = EMPTY_MESSAGE)
-    @Pattern(regexp = "^[-_!?.,a-zA-Z0-9가-힣ㄱ-ㅎㅏ-ㅣ]{1,20}$", message = ORGANIZATION_MESSAGE)
-    private String organization;
-
-    public MemberSaveRequest() {
+    public LoginRequest() {
     }
 
-    public MemberSaveRequest(final String email, final String password, final String organization) {
+    public LoginRequest(final String email, final String password) {
         this.email = email;
         this.password = password;
-        this.organization = organization;
     }
 
     public String getEmail() {
@@ -35,9 +30,5 @@ public class MemberSaveRequest {
 
     public String getPassword() {
         return password;
-    }
-
-    public String getOrganization() {
-        return organization;
     }
 }
