+++ backend/src/main/java/com/woowacourse/zzimkkong/dto/reservation/ReservationDeleteRequest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.zzimkkong.dto;
+package com.woowacourse.zzimkkong.dto.reservation;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
