+++ backend/src/main/java/com/woowacourse/zzimkkong/service/AuthService.java
@@ -1,8 +1,8 @@
 package com.woowacourse.zzimkkong.service;
 
 import com.woowacourse.zzimkkong.domain.Member;
-import com.woowacourse.zzimkkong.dto.LoginRequest;
-import com.woowacourse.zzimkkong.dto.TokenResponse;
+import com.woowacourse.zzimkkong.dto.member.LoginRequest;
+import com.woowacourse.zzimkkong.dto.member.TokenResponse;
 import com.woowacourse.zzimkkong.exception.NoSuchMemberException;
 import com.woowacourse.zzimkkong.exception.PasswordMismatchException;
 import com.woowacourse.zzimkkong.infrastructure.JwtUtils;
