+++ backend/src/main/java/com/woowacourse/zzimkkong/service/MemberService.java
@@ -1,8 +1,8 @@
 package com.woowacourse.zzimkkong.service;
 
 import com.woowacourse.zzimkkong.domain.Member;
-import com.woowacourse.zzimkkong.dto.MemberSaveRequest;
-import com.woowacourse.zzimkkong.dto.MemberSaveResponse;
+import com.woowacourse.zzimkkong.dto.member.MemberSaveRequest;
+import com.woowacourse.zzimkkong.dto.member.MemberSaveResponse;
 import com.woowacourse.zzimkkong.exception.DuplicateEmailException;
 import com.woowacourse.zzimkkong.repository.MemberRepository;
 import org.springframework.stereotype.Service;
